Import changes needed for CUDA-NVLINK support
There are multiple things in this changeset:

   - Add a lock in the data_t to enable multiple GPU drivers to
     organize mutual exclusion wrt metadata update

   - upgrade the data_key_t to 64 bits to handle larger sets of
     keys

   - extend list.h and zone_malloc with a few utility functions

   - split data_transfer_ownership_to_copy in two stages

   - rewrite dev_cuda to manage NVLINK and allow to pull read-only
     copies from other devices instead of systematically from the
     RAM

   - Fix asymetric change in management of profiling for GPU tasks (it's still unclear how a better way would work).

   - Fix warnings and compilation errors when enabling debug, paranoid, noiser, trace, prof_graph, etc.

+++ parsec/data_dist/matrix/sym_two_dim_rectangle_cyclic_band.c
@@ -27,8 +27,8 @@ static inline void sym_twoDBC_band_offset(sym_two_dim_block_cyclic_t *dc,
     *m += dc->super.i / dc->super.mb;
     *n += dc->super.j / dc->super.nb;
 
-    assert( *m < dc->super.mt );
-    assert( *n < dc->super.nt );
+    assert( *m < (unsigned int)dc->super.mt );
+    assert( *n < (unsigned int)dc->super.nt );
 }
 
 /* New rank_of for sym two dim block cyclic band */
@@ -49,7 +49,7 @@ uint32_t sym_twoDBC_band_rank_of(parsec_data_collection_t * desc, ...)
             (dc->super.uplo == matrix_Upper && n>=m) );
 
     /* Check tile location */
-    if( abs(m-n) < dc->band_size ){
+    if( (unsigned int)abs(m-n) < dc->band_size ){
         /* New index */
         m = abs(m - n); 
 
@@ -97,7 +97,7 @@ parsec_data_t* sym_twoDBC_band_data_of(parsec_data_collection_t *desc, ...)
     key = (n * dc->super.super.lmt) + m;
 
     /* Check tile location */
-    if( abs(m-n) < dc->band_size ){
+    if( (unsigned int)abs(m-n) < dc->band_size ){
         /* The new m in band */
         m = abs(m - n);
